aider: Added check to ensure that the target directory exists before proceeding with the rest of the main function.
# Aider chat conversation:

ASSISTANT: To enforce strict parsing of the TOML file, ensuring that it fails if there are any fields not defined in the struct, you can use the `#[serde(deny_unknown_fields)]` attribute on your structs. This attribute will cause `serde` to return an error during deserialization if it encounters any fields in the input that are not specified in the struct.

You'll need to add this attribute to the `Config`, `Program`, and `Pdp11` structs in the `pargo/src/conf.rs` file.

Here's how you can modify the `Config`, `Program`, and `Pdp11` structs:

```rust
pargo/src/conf.rs
<<<<<<< SEARCH
#[derive(Serialize, Deserialize, Debug)]
pub struct Config {
=======
#[derive(Serialize, Deserialize, Debug)]
#[serde(deny_unknown_fields)]
pub struct Config {
>>>>>>> REPLACE
    pub program: Program,
    pub pdp11: Option<Pdp11>,
}

<<<<<<< SEARCH
#[derive(Serialize, Deserialize, Debug)]
pub struct Program {
=======
#[derive(Serialize, Deserialize, Debug)]
#[serde(deny_unknown_fields)]
pub struct Program {
>>>>>>> REPLACE
    pub name: String,
    pub version: String,
    pub platform: Platform,
}

<<<<<<< SEARCH
#[derive(Serialize, Deserialize, Debug)]
pub struct Pdp11 {}
=======
#[derive(Serialize, Deserialize, Debug)]
#[serde(deny_unknown_fields)]
pub struct Pdp11 {}
>>>>>>> REPLACE
```

By adding the `#[serde(deny_unknown_fields)]` attribute, you instruct `serde` to ensure that all fields in the TOML file have corresponding fields in the Rust structs and to fail otherwise.

USER: InvalidEditBlock: edit failed!

} does not contain the *exact chunk* of SEARCH lines you specified.
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The SEARCH block needs to be EXACTLY the same as the lines in } with nothing missing!

} does not contain these 2 exact lines in a row:
```
#[derive(Serialize, Deserialize, Debug)]
pub struct Program {
```

USER: In main.rs 'let target_dir = project_path.join("target");' I locate a path of a directory. I'd like to ensure that this directory exists (i.e. create it if it doesn't exist).

ASSISTANT: To ensure that the `target` directory exists, you can use the `std::fs::create_dir_all` function, which recursively creates a directory and all of its parent components if they are missing. You'll need to modify the `main` function in the `pargo/src/main.rs` file to include this check and creation step.

Here's the change you need to make:

```rust
pargo/src/main.rs
<<<<<<< SEARCH
    let target_dir = project_path.join("target");
=======
    let target_dir = project_path.join("target");
    std::fs::create_dir_all(&target_dir)?;
>>>>>>> REPLACE
```

This change ensures that before proceeding with the rest of the `main` function, the `target` directory is created if it does not already exist. If there's an error creating the directory, it will be propagated up by the `?` operator, and the function will return an `Error`.